+++ tests/benchmarks/igraph_random_walk.c
@@ -24,37 +24,51 @@ int main() {
         VECTOR(weights)[i] = igraph_rng_get_unif01(igraph_rng_default());
     }
 
+    /* Only edges */
+
     BENCH(" 1 Random edge walk,   directed,   unweighted, small graph  ",
-          igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+          igraph_random_walk(&graph, NULL, NULL, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
          );
 
     BENCH(" 2 Random edge walk,   directed,   weighted,   small graph  ",
-          igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+          igraph_random_walk(&graph, &weights, NULL, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
          );
 
-    BENCH(" 3 Random vertex walk, directed,   unweighted, small graph  ",
-          igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    BENCH(" 3 Random edge walk,   undirected, unweighted, small graph  ",
+        igraph_random_walk(&graph, NULL, NULL, &edges, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    );
+
+    BENCH(" 4 Random edge walk,   undirected, weighted,   small graph  ",
+        igraph_random_walk(&graph, &weights, NULL, &edges, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    );
+
+    /* Only vertices */
+
+    BENCH(" 5 Random vertex walk, directed,   unweighted, small graph  ",
+          igraph_random_walk(&graph, NULL, &vertices, NULL, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
          );
 
-    BENCH(" 4 Random vertex walk, directed,   weighted, small graph  ",
-          igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    BENCH(" 6 Random vertex walk, directed,   weighted, small graph  ",
+          igraph_random_walk(&graph, &weights, &vertices, NULL, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
          );
-
-    BENCH(" 5 Random edge walk,   undirected, unweighted, small graph  ",
-          igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
-         );
-
-    BENCH(" 6 Random edge walk,   undirected, weighted,   small graph  ",
-          igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
-         );
-
+    
     BENCH(" 7 Random vertex walk, undirected, unweighted, small graph  ",
-          igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+          igraph_random_walk(&graph, NULL, &vertices, NULL, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
          );
 
     BENCH(" 8 Random vertex walk, undirected, weighted, small graph  ",
-          igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+          igraph_random_walk(&graph, &weights, &vertices, NULL, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
          );
+
+    /* Both edges and vertices */
+
+    BENCH(" 9 Random walk,   directed,   unweighted, small graph  ",
+        igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    );
+
+    BENCH(" 10 Random walk,   directed,   weighted,   small graph  ",
+        igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    );
 
     igraph_destroy(&graph);
 
@@ -67,69 +81,97 @@ int main() {
         VECTOR(weights)[i] = igraph_rng_get_unif01(igraph_rng_default());
     }
 
-    BENCH(" 9 Random edge walk,   directed,   unweighted, large graph  ",
-          igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    /* Only edges */
+
+    BENCH(" 11 Random edge walk,   directed,   unweighted, large graph  ",
+          igraph_random_walk(&graph, NULL, NULL, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
          );
 
-    BENCH(" 10 Random edge walk,   directed,   weighted,   large graph  ",
-          igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    BENCH(" 12 Random edge walk,   directed,   weighted,   large graph  ",
+          igraph_random_walk(&graph, &weights, NULL, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
          );
 
-    BENCH(" 11 Random vertex walk, directed,   unweighted, large graph  ",
-          igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
-         );
-
-    BENCH(" 12 Random vertex walk, directed,   weighted, large graph  ",
-          igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
-         );
-    
     BENCH(" 13 Random edge walk,   undirected, unweighted, large graph  ",
-          igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
-         );
+        igraph_random_walk(&graph, NULL, NULL, &edges, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    );
 
     BENCH(" 14 Random edge walk,   undirected, weighted,   large graph  ",
-          igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+        igraph_random_walk(&graph, &weights, NULL, &edges, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    );
+
+    /* Only vertices */
+
+    BENCH(" 15 Random vertex walk, directed,   unweighted, large graph  ",
+          igraph_random_walk(&graph, NULL, &vertices, NULL, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
          );
 
-    BENCH(" 15 Random vertex walk, undirected, unweighted, large graph  ",
-          igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    BENCH(" 16 Random vertex walk, directed,   weighted, large graph  ",
+          igraph_random_walk(&graph, &weights, &vertices, NULL, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
          );
 
-    BENCH(" 16 Random vertex walk, undirected, weighted, large graph  ",
-          igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    BENCH(" 17 Random vertex walk, undirected, unweighted, large graph  ",
+          igraph_random_walk(&graph, NULL, &vertices, NULL, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
          );
 
-    BENCH(" 17 Short edge walk,    directed,   unweighted, large graph, x 100",
-          REPEAT(igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_OUT, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
+    BENCH(" 18 Random vertex walk, undirected, weighted, large graph  ",
+          igraph_random_walk(&graph, &weights, &vertices, NULL, 0, IGRAPH_ALL, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
          );
 
-    BENCH(" 18 Short edge walk,    directed,   weighted,   large graph, x 100",
-          REPEAT(igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_OUT, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
+    /* Both edges and vertices */
+
+    BENCH(" 19 Random walk,   directed,   unweighted, large graph  ",
+        igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    );
+
+    BENCH(" 20 Random walk,   directed,   weighted,   large graph  ",
+        igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_OUT, 50000000, IGRAPH_RANDOM_WALK_STUCK_RETURN)
+    );
+
+    /* Only edges */
+
+    BENCH(" 21 Short edge walk,    directed,   unweighted, large graph, x 100",
+          REPEAT(igraph_random_walk(&graph, NULL, NULL, &edges, 0, IGRAPH_OUT, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
          );
 
-    BENCH(" 19 Short vertex walk,    directed,   unweighted, large graph, x 100",
-          REPEAT(igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_OUT, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
+    BENCH(" 22 Short edge walk,    directed,   weighted,   large graph, x 100",
+          REPEAT(igraph_random_walk(&graph, &weights, NULL, &edges, 0, IGRAPH_OUT, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
          );
 
-    BENCH(" 20 Short vertex walk,    directed,   weighted,   large graph, x 100",
-          REPEAT(igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_OUT, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
+    BENCH(" 23 Short edge walk,    undirected, unweighted, large graph, x 100",
+        REPEAT(igraph_random_walk(&graph, NULL, NULL, &edges, 0, IGRAPH_ALL, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
+    );
+
+    BENCH(" 24 Short edge walk,    undirected, weighted,   large graph, x 100",
+        REPEAT(igraph_random_walk(&graph, &weights, NULL, &edges, 0, IGRAPH_ALL, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
+    );
+
+    /* Only vertices */
+
+    BENCH(" 25 Short vertex walk,    directed,   unweighted, large graph, x 100",
+          REPEAT(igraph_random_walk(&graph, NULL, &vertices, NULL, 0, IGRAPH_OUT, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
          );
 
-    BENCH(" 21 Short edge walk,    undirected, unweighted, large graph, x 100",
-          REPEAT(igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_ALL, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
-         );
-
-    BENCH(" 22 Short edge walk,    undirected, weighted,   large graph, x 100",
-          REPEAT(igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_ALL, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
+    BENCH(" 26 Short vertex walk,    directed,   weighted,   large graph, x 100",
+          REPEAT(igraph_random_walk(&graph, &weights, &vertices, NULL, 0, IGRAPH_OUT, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
          );
     
-    BENCH(" 23 Short vertex walk,    undirected, unweighted, large graph, x 100",
-          REPEAT(igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_ALL, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
+    BENCH(" 27 Short vertex walk,    undirected, unweighted, large graph, x 100",
+          REPEAT(igraph_random_walk(&graph, NULL, &vertices, NULL, 0, IGRAPH_ALL, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
          );
     
-    BENCH(" 24 Short vertex walk,    undirected, weighted,   large graph, x 100",
-          REPEAT(igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_ALL, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
+    BENCH(" 28 Short vertex walk,    undirected, weighted,   large graph, x 100",
+          REPEAT(igraph_random_walk(&graph, &weights, &vertices, NULL, 0, IGRAPH_ALL, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
          );
+
+    /* Both edges and vertices */
+
+    BENCH(" 29 Short random walk,    directed,   unweighted, large graph, x 100",
+        REPEAT(igraph_random_walk(&graph, NULL, &vertices, &edges, 0, IGRAPH_OUT, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
+    );
+
+    BENCH(" 30 Short random walk,    directed,   weighted,   large graph, x 100",
+        REPEAT(igraph_random_walk(&graph, &weights, &vertices, &edges, 0, IGRAPH_OUT, 10000, IGRAPH_RANDOM_WALK_STUCK_RETURN), 100)
+    );
 
     igraph_destroy(&graph);